pyproject.toml
[project]
name = "monzo-tracker"
version = "0.1.0"
description = "A comprehensive financial tracker using the Monzo API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]

[project.scripts]
monzo-tracker = "monzo_tracker.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monzo_tracker"]
